requirements.txt
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
dbt-postgres==1.7.2
apache-airflow==2.7.1
python-dotenv==1.0.0
